start kicking players early when restarting
Players begin to be kicked at 4 seconds until server close (1 per tick).

All online players are kicked and the cache/every_tick/main function is also called before the server closes so that player cache data is cleaned up in case of a change in logic after the data pack is reloaded.