refactored stored item db recycling
stored item IDs are no longer recycled, only the slots. The entries are removed from the database when they are recycled, the slot information is appended to a new list "reuse_slots". When a new stored item is made, the slot info from that list is prioritised and a new entry is always created in the db.